+++ test/Components/FragmentTest.php
@@ -54,7 +54,6 @@ class FragmentTest extends PHPUnit_Framework_TestCase
         return [
             'bool'      => [true],
             'Std Class' => [(object) 'foo'],
-            'null'      => [null],
             'float'     => [1.2],
             'array'      => [['foo']],
         ];