+++ test/Components/QueryTest.php
@@ -36,7 +36,6 @@ class QueryTest extends PHPUnit_Framework_TestCase
         return [
             'bool'      => [true],
             'Std Class' => [(object) 'foo'],
-            'null'      => [null],
             'float'     => [1.2],
             'array'      => [['foo']],
         ];